Invoice: move invoice parsing code to a specific non-gui module
This moves some code related to invoice formats to a specific module.
The behavior of the GUI is mostly unchanged, except it is a bit more strict about the input file: address is now mandatory in an invoice and no exchange rate can be specified when currency is XEC.

Various setters and getters now expect/return more precise types (Address or Decimal instead of a generic string).