fix ColorScheme for dark style
The second call to `ColorScheme.update_from_widget` did not take the user setting  (`force_dark`) into consideration, so it overwrote the ColorScheme setting. This is strange because I would expect the first call to change the stylesheet, and so make the second call autodetect that the dark theme is in use via `ColorScheme.has_dark_background`, but it somehow does not work. Maybe Qt change something in the way the stylesheet is applied, and it is not taken into account by the time the second call occurs.

This changes the behavior to the one used by electrum. Hopefully there aren't too many Electrum ABC plugins messing with the color theme.