restrict protobuf library version to <= 3.20.1
There are backwards incompatible changes in the latest release of protobuf. See #214 and https://developers.google.com/protocol-buffers/docs/news/2022-05-06#python-updates

This is a temporary fix by setting an upper limit to protobuf library.

The proper fix would be to regenerate the _pb2.py files, however it is unclear at the moment if this would keep working on systems with `protobuf<4.21`. For reference, ubuntu 22.04 currently has protobuf version 3.12.4.