try pycryptodome as fallback implementation of ripemd160
Before defaulting to the python implementation of ripemd, try pycryptodome. It is about 25 times faster.

Note that for now it is not added as a direct dependency in the requirements files, except in the deterministic requirements. It will usually be available for most forms of installs,  because it is an indirect dependency (via one of the direct dependencies).