Fix to blockchain.py get_bits (was sometimes wrong)
This function sometimes returned a target larger than MAX_TARGET,
which could break theoretically on any testnet
when difficulty would drop to insanely low values (below 1).

This has been addressed.

Also in this commit: We re-wrote the target_to_bits and bits_to_target
functions to exactly copy the logic and code in  the node's arith_uint256.cpp,
so as to avoid any potential bugs with respect to the implementation. These
two functions are pretty much a direct translation now of the C++
arith_uint256::SetCompact() and arith_uint256::GetCompact() functions.