avalanche proofs: put the block height of a UTXO in orange if not enough confirmations
A UTXO must have 2016 confirmations to be used in a valid proof. Put the block height in orange and explain why  in the tooltip.
The reason for using orange instead of red is that the proof will eventually become valid. Users might know what they are doing when they build the proof in advance of it being usable.