[proof editor] store SignedStake or Stake + Key internally
This removes boilerplate code attempting to analyze the dict objects used previously. Now, the dict object will only be used initially when loading coins from a file or building from coins in this wallet, and it will be converted internally into a Stake + Key (we need to wait for the rest of the proof data to be final before signing the stake).

Remove also the coin freezing feature on proof building. The coins will likely not belong to the wallet building the proof.