[Proof builder] enable building proof without master private key (with invalid sig)
This changes the proof builder to make it accept a master pub instead of a master private key.
The resulting signature will be invalid.

This is to allow people to add signed stakes to a proof without knowledge of the master key. Such an invalid proof can be used to gather stakes from multiple users, and merge all the proofs in the master wallet (who can then produce a correct signature).