[flake8] run flake8-mutable extension to all files
Python's default arguments are evaluated at definition as opposed to when the function is invoked. This leads to unexpected behavior, as mutations persist between calls. For a more detailed explanation, see https://docs.python-guide.org/writing/gotchas/#mutable-default-arguments

This is rarely a good idea, and can cause nasty and hard to debug issues when done accidentaly. Don't allow it.