[main_window] move imports at the top of the file
This is a PEP8 rule:
> Imports are always put at the top of the file, just after any module comments and docstrings, and before module globals and constants.

This fixes a lot of the  E402 flake8 errors. None of these imports had a good reason to not be at the top of the file (lazy loading, circular dependencies...)

Test by running the program and checking it does not fail due to circular import, then run `flake8 --select=E402 electroncash_gui/mainwindow.py`

Related to #252.