[CI] run test suite on Windows; comment out coverage report
Disable `coverage report` which no longer works on cirrus for an unknown reason and causes the check to appear as failed even though all the tests pass.

Simplify the script to make sure it works on all OS (`set -e` does not work on windows, you can't use `;` at the end of a command on Windows...)

Remove a few tasks when compiling secp256k1. We don't need to re-test all versions of python for this, only the most recent.