Revert "Build: Verify the Python checksums"
This reverts commit 059af1127a8976c797040427d20d922fe142cfa2.

The motivation for this feature was:

> Before this change, whoever has a copy of the Python release signing key
> is a trusted third party. This commit changes this so that in addition
> to checking the signature of the installers, the checksums are also
> verified.

The problem is that makes it more difficult to bump the python version used for packaging the application. It is not just a matter of downloading the msi files and running `sha256sum` on them, there is also the cost of forgetting to do it, wasting an hour on running the build script and then spending time troubleshooting the issue (which is not immediately trivial if you have not encountered it before).

It seems reasonable to assume that whoever has the Python release signing key is a trusted third party for a python program.