[flake8] silence E722 errors (do not use bare except)
I did this semi-automatically by just replacing bare exceptions with `except Exception`. It is still awful, but at least we no longer catch `KeyboardInterrupt` and `SystemExit`.

Fixing this properly would take an enormous amount of work, understanding each one of the 100+ occurences to guess the correct exceptions to catch. This will have to be done the slow way, each occurence at a time when the code is touched for another reason.