don't inherit, raise or catch BaseException
BaseException is the base class for all builtin exceptions, including stuff like SystemExit and KeyboardInterrupt. It is rarely a good idea to catch those. This is equivalent to bare except clauses (see next commit)
The official python documentation explicitely says to not directly inherit this for custom exceptions, but inherit Exception instead.

When it comes to explicitely raising an error, RuntimeError is generic enough and can be caught with a generic `except Exception`.