move settings dialog code to its own module
This move 900 terrible lines of code out of main_window.py and into its own module. The whole thing is still very monolithic and coupled with the main window, but at least now the coupling is visible in the __init__ parameters and connections, and does not require to read the whole code any longer.