changing unit requires to restart the application
Some things are automatically updated, such as the displayed amounts in the tx history, status bar ... But the amount input fields (e.g. in the send tab) require a restart to be updated.
I don't know if changing the unit should be supported forever, so I'm not going to spend much effort on being able to dynamically change the unit everywhere in the application.