address consolidation: fix min/max value/height selection
The address consolidation tool did not seem to work on testnet. The problem was that we always pass a max block height even if the corresponding checkbox is not checked, and that the default value is lower than the current testnet block height.

This fixes the issue by passing None if the checkbox is unchecked, and also by setting a larger default maximum block height (in case users check the box but don't increase the value). This should give us some wiggle room.

The problem is the same for min and max value, and min block height, except that the default values for these are more sane (less likely to filter out values by accident)