fix sys.platform detection
In this instances, the parentheses indicate that the programmer's  intention is to check if the value is one of a tuple of strings. However, in python the syntax for a tuple of one element requires a trailing comma. Without the comma, the parentheses are considered redudant and we are actually checking if the value is a substring of the string between parentheses.

In this case, it probably worked anyway, because "win32" happens to be a substring of "win32", and "linux" a substring of "linux". But better to fix it.

Note that these platform detection code snippets are inconsistent throughout the codebase. Sometimes we use `sys.platform in ("windows", "win32")`, sometimes we use `sys.platform in ("win32")`, and sometimes it is `sys.platform in ("win32", "win64", "windows")`. I'm not sure if there is a good reason for that, nor if `win64` and `windows` are actually possible values for `sys.platform` (the official documentation only mentions `win32` and `cygwin`).