fix portable behavior
The main change here is that we fix the broken portable.exe behavior caused by PyInstaller now setting the `__file__` attribute to the temp folder where the bundle is temporarily extracted. We don't want the data dir in a temp folder, we want it next to the exe.

As a bonus, this probably fixes the `--portable` CLI option which I suspect was broken in Electron Cash when an attempt was made to support py2app. This will allow to run the AppImage or source dist with that option, from a hard drive or USB thumbdrive.

This also enables deprecation warnings for developers (anyone running this from a git clone)

----

revert py2app support chenge for portable detection

https://github.com/Electron-Cash/Electron-Cash/pull/345/commits/c99d4e213ee1a0eaefbcf4d9195c1136a0d1d1e2

----

 run_electrum: small clean-up, and hide DeprecationWarnings if not git

- rename `is_bundle` to `is_pyinstaller` (no semantic changes, just to clearer name)
- define `is_appimage`
- add comment to explain `is_local`
  - its value is the same as before (but more explicit definition)
- define `is_git_clone`, and restrict DeprecationWarnings to that case

https://github.com/spesmilo/electrum/commit/5f1a13e4ea9bd3b1a1780e3f7888ea1e9af0765c

----

"--portable": make behaviour independent of pyinstaller version

pyinstaller 4.3 changed the value of `__file__`.
This change makes our behaviour independent of that pyinstaller change
(we always behave like old versions of pyinstaller).

https://github.com/spesmilo/electrum/commit/5b500f08ea2618c08b47a99176f323810c924690

----

"--portable": more consistent behaviour

The old and new behaviour is as follows:
1. "pyinstaller" case: portable `.exe`, other `.exe`s with `--portable`, and `.dmg` with `--portable`
    - uses `$PWD`
    - note that when you double-click the portable `.exe` on Windows, `$PWD` is set to the parent folder, i.e. the datadir gets put next to the `.exe`
2. appimage `--portable`
    - was broken (see #5551)
    - (CHANGED NOW to) uses `$PWD`
3. git clone
    - next to `run_electrum`
4. unpacking `tar.gz` and running locally from it
    - next to `run_electrum`
5. `pip install *.tar.gz`, and calling "electrum --portable" from terminal
    - used python's user script directory
        - `~/.local/bin/electrum_data`
        - `$VIRTUAL_ENV/bin/electrum_data`
    - (CHANGED NOW to) uses `$PWD`

That is, we now almost always put the datadir in `$PWD`,
except for the local source case, where we put it next to `run_electrum`.

The "appimage" case (2) is now fixed.

The only breaking change is re case 5 which previously behaved completely
unintuitively and most likely not in a useful way.

https://github.com/spesmilo/electrum/commit/6a34d93ce2388feeae320ab1c614bc48ad2d4b31