fix integer overflow for amount selection
Qt translates `int` into a C int (32 bytes signed integer), which causes overflows when the selected amount in satoshis is larger than the int32 max.

The tradeoff of using a generic PythonObject is performance, as there will be a lot of overhead. But this is not an issue as this codepath is only triggered rarely and at low frequency, by users cliocking on coins or addresses in the GUI.