add a checkbox to disable shuffling the outputs when using op_return
The intended use case for this is sending SLP token using a hex encoded op_return payload.

Test plan:
Check that the new checkbox is only enabled if something is typed into the op_return payload field, and disabled when this payload is deleted.

Add multiple outputs in the send tab, leave the op_return field blank. Click "Preview" multiple times and make sure the outputs are shuffled (no matter if the new checkbox is checked or not).

Type something in the op_return field, and repeat the above test. Now the shuffling behavior must depend on the checkbox. The op_return output should always be first. The other outputs should be shuffled if the checkbox is checked. If it is unchecked, then the change output must be last, and preceded by the other outputs in the specified order.

Click send and check the order in a block explorer.